backend/src/PaymentAgent/requirements.txt
stripe>=8.0.0
google-generativeai>=0.4.0
playwright>=1.40.0
aiohttp>=3.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.109.0
httpx>=0.26.0
